# docs/api/openapi.yaml
components:
  schemas:
...
    ApplicationResponse:
      properties:
        amountMinorUnits:
          description: "The requested principal in the currency's MINOR units - cents, not euros. Money is never a floating-point number here."
          example: 1500000
          format: int64
          type: integer
        applicantInitials:
          description: "Initials only. The API never returns an applicant's name, not even to the client that supplied it."
          example: T.A.
          type: string
        applicantReference:
          description: |
            An irreversible HMAC pseudonym, stable for the same applicant across applications.

            **Pseudonymised, not anonymised**: the same applicant always produces the same value, so it is linkable and remains personal data under GDPR.
          example: APR-fa05675c49b42435
          pattern: "^APR-[0-9a-f]{16}$"
          type: string
        applicationId:
          description: "Opaque, server-generated."
          format: uuid
          type: string
        createdAt:
          format: date-time
          type: string
        currency:
          type: string
        decision:
          $ref: "#/components/schemas/Decision"
          description: Null until a decision has been recorded.
        productCode:
          type: string
        purpose:
          type: string
        residenceCountry:
          type: string
        status:
          description: The application's position in its lifecycle.
          enum:
          - DRAFT
          - SUBMITTED
          - VALIDATING
          - CHECKS_IN_PROGRESS
          - MANUAL_REVIEW
          - APPROVED
          - REJECTED
          - FAILED
          - CANCELLED
          type: string
        submittedAt:
          format: date-time
          type: string
        termMonths:
          format: int32
          type: integer
        updatedAt:
          format: date-time
          type: string
        version:
          description: Optimistic-locking version.
          format: int64
          type: integer
      type: object
...
    Decision:
      description: "The recorded outcome, once there is one."
      properties:
        decidedBy:
          description: "The deciding reviewer's pseudonymous token subject, or null when the decision was automatic. Never a name."
          type: string
        reasonCode:
          description: "A stable code, never prose. Free text about why a named individual was refused credit is exactly what must not travel through events into the audit store."
          example: CREDIT_SCORE_BELOW_THRESHOLD
          type: string
        type:
          description: |
            `MANUAL_REVIEW` means the platform declined to decide. That is not the same as deciding against the applicant, and the record has to be able to tell the two apart later.
          enum:
          - APPROVED
          - REJECTED
          - MANUAL_REVIEW
          type: string
      type: object
...
    PageResponseApplicationResponse:
      properties:
        hasNext:
          type: boolean
        items:
          items:
            $ref: "#/components/schemas/ApplicationResponse"
          type: array
        page:
          format: int32
          type: integer
        size:
          format: int32
          type: integer
        totalItems:
          format: int64
          type: integer
        totalPages:
          format: int32
          type: integer
      type: object
    Problem:
      description: |
        An RFC 9457 Problem Detail.

        It never contains a stack trace, a SQL fragment, a table name, an internal identifier, or an exception message. Validation errors name the **field** and never echo its value, because the value is frequently the personal data the platform exists to protect.

        Branch on `errorCode`, never on `detail`.
      properties:
        correlationId:
          description: "Ties this response to the platform's logs. Quote it when reporting a problem: it is deliberately the only link, because the body carries nothing else."
          type: string
        detail:
          description: Prose for a human. May be reworded between releases.
          type: string
        errorCode:
          description: Stable and machine-readable. This is what a client branches on.
          enum:
          - REQUEST_VALIDATION_FAILED
          - MALFORMED_REQUEST_BODY
          - INVALID_REQUEST_VALUE
          - INVALID_PATH_PARAMETER
          - ACCESS_DENIED
          - APPLICATION_NOT_FOUND
          - APPLICATION_NOT_EDITABLE
          - DOCUMENT_NOT_FOUND
          - ILLEGAL_STATE_TRANSITION
          - ILLEGAL_DOCUMENT_STATE
          - MANDATORY_DOCUMENTS_MISSING
          - IDEMPOTENCY_KEY_REUSED
          - IDEMPOTENCY_KEY_IN_USE
          - CONCURRENT_MODIFICATION
          - CONCURRENT_DOCUMENT_UPDATE
          - OBJECT_STORAGE_UNAVAILABLE
          - SCANNER_UNAVAILABLE
          - INTERNAL_ERROR
          type: string
        instance:
          example: /v1/applications/018f3a9c-6b21-7c4e-9d55-2a1b8e7f0c33/submit
          type: string
        status:
          example: 409
          format: int32
          type: integer
        timestamp:
          format: date-time
          type: string
        title:
          example: Conflict
          type: string
        type:
          example: about:blank
          type: string
      type: object
    ReviewDecisionRequest:
      properties:
        decision:
          minLength: 1
          pattern: ^(APPROVED|REJECTED)$
          type: string
      required:
      - decision
      type: object
...
  securitySchemes:
    oauth2:
      description: |
        An OAuth2 bearer token, validated at API Gateway **and again by each service**. The second validation is what protects the services if anything is ever deployed with a route that bypasses the gateway.

        Partners additionally present an API key. The key is what a usage plan throttles on — a capacity identity, never an authentication one — so a key alone grants nothing.
      flows:
        clientCredentials:
          scopes:
            applications:read: Read applications and their status
            applications:write: "Create, update and submit applications"
            documents:read: List documents on an application
            documents:write: Request upload slots and confirm uploads
            manual-review:decide: Record a reviewer's decision
            manual-review:read: Read the manual review queue
          tokenUrl: https://auth.example.com/oauth2/token
      type: oauth2
info:
  contact:
    name: Repository
    url: https://example.com/aws-loan-origination-platform
  description: |
    ## What this is

    A **reference implementation**. It has never been deployed, it has never processed real data, and every external check behind it — KYC, AML, fraud, credit scoring, malware scanning — is a **simulation** that contacts no real provider.

    ## Two services, one API surface

    These paths are served by two independently deployed services behind a single API Gateway. `/v1/applications/{applicationId}/documents/**` belongs to the document service; everything else belongs to the application service. Callers should not depend on that split — it is recorded because it explains why a document can be accepted a moment before submission will accept it.

    ## Idempotency

    `POST /v1/applications` and `POST .../submit` accept an `Idempotency-Key` header, and submission is the one that most needs it: a client that times out and retries must not submit twice.

    | Situation | Result |
    |---|---|
    | Same key, same request | The original response, with `Idempotent-Replay: true` |
    | Same key, different request | `409` with `IDEMPOTENCY_KEY_REUSED` |
    | Same key, original still in flight | `409` with `IDEMPOTENCY_KEY_IN_USE` |

...

    ## Errors

...

    Branch on `errorCode`, never on `detail`, which is prose and may be reworded.

    ## Personal data

...

...
  license:
    identifier: Apache-2.0
    name: Apache-2.0
...
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/manual-review/tasks:
    get:
      description: |
        Oldest first, paginated and bounded. An unbounded queue endpoint is a denial-of-service vector against the platform's own database as much as against the caller.
      operationId: listTasks
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 0
          format: int32
          type: integer
      - in: query
        name: size
        required: false
        schema:
          default: 20
          format: int32
          type: integer
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseApplicationResponse"
          description: A page of applications awaiting review.
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: The token lacks `manual-review:read`.
      summary: Applications awaiting a human decision
      tags:
      - Manual review
  /v1/manual-review/tasks/{taskId}/decisions:
    post:
      description: |
        The decision is attributed to the **token's subject**, never to a value the client sends in the body: a caller must not be able to attribute a decision to someone else.

        There is deliberately no free-text notes field. Reviewer notes would flow into the decision record and from there into events and the audit store, where free text about a named individual is precisely what the platform's privacy rules exclude.
      operationId: decide
      parameters:
      - description: "The application under review. The queue is a projection of applications, so a task identifier IS an application identifier."
        in: path
        name: taskId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecisionRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
          description: The application after the decision.
        "403":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: The token lacks `manual-review:decide`.
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Problem"
          description: The application is no longer awaiting review.
      summary: Record a reviewer's decision
      tags:
      - Manual review
security:
- oauth2: []
servers:
...